# webapi.event+/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Event+",
    "description": "API para gereciamento de Eventos - Backend API",
    "contact": {
      "name": "Joao Victor Alves Bautista - Turma Manhã",
      "url": "https://github.com/zAlves31"
    },
    "version": "v1"
  },
  "paths": {
    "/api/ComentariosEvento": {
      "get": {
        "tags": [
          "ComentariosEvento"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ComentariosEvento"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ComentariosEvento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ComentariosEvento/BuscarPorIdUsuario": {
      "get": {
        "tags": [
          "ComentariosEvento"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idUsuario",
            "type": "string",
            "format": "uuid",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "idEvento",
            "type": "string",
            "format": "uuid",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ComentariosEvento/{id}": {
      "delete": {
        "tags": [
          "ComentariosEvento"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Evento/{id}": {
      "get": {
        "tags": [
          "Evento"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Evento"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Evento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Evento"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ComentariosEvento": {
      "required": [
        "descricao",
        "exibe",
        "idEvento",
        "idUsuario"
      ],
      "type": "object",
      "properties": {
        "idComentarioEvento": {
          "format": "uuid",
          "type": "string"
        },
        "descricao": {
          "minLength": 1,
          "type": "string"
        },
        "exibe": {
          "type": "boolean"
        },
        "idUsuario": {
          "format": "uuid",
          "type": "string"
        },
        "usuario": {
          "$ref": "#/definitions/Usuario"
        },
        "idEvento": {
          "format": "uuid",
          "type": "string"
        },
        "evento": {
          "$ref": "#/definitions/Evento"
        }
      },
      "additionalProperties": false
    },
    "Evento": {
      "required": [
        "dataEvento",
        "descricao",
        "nomeEvento"
      ],
      "type": "object",
      "properties": {
        "idEvento": {
          "format": "uuid",
          "type": "string"
        },
        "dataEvento": {
          "format": "date-time",
          "type": "string"
        },
        "nomeEvento": {
          "minLength": 1,
          "type": "string"
        },
        "descricao": {
          "minLength": 1,
          "type": "string"
        },
        "idTipoEvento": {
          "format": "uuid",
          "type": "string"
        },
        "tiposEvento": {
          "$ref": "#/definitions/TiposEvento"
        },
        "idInstituicao": {
          "format": "uuid",
          "type": "string"
        },
        "instituicao": {
          "$ref": "#/definitions/Instituicao"
        }
      },
      "additionalProperties": false
    },
    "Instituicao": {
      "required": [
        "cnpj",
        "endereco",
        "nomeFantasia"
      ],
      "type": "object",
      "properties": {
        "idInstituicao": {
          "format": "uuid",
          "type": "string"
        },
        "cnpj": {
          "maxLength": 14,
          "minLength": 0,
          "type": "string"
        },
        "endereco": {
          "minLength": 1,
          "type": "string"
        },
        "nomeFantasia": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "TiposEvento": {
      "required": [
        "titulo"
      ],
      "type": "object",
      "properties": {
        "idTipoEvento": {
          "format": "uuid",
          "type": "string"
        },
        "titulo": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "TiposUsuario": {
      "required": [
        "titulo"
      ],
      "type": "object",
      "properties": {
        "idTipoUsuario": {
          "format": "uuid",
          "type": "string"
        },
        "titulo": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Usuario": {
      "required": [
        "email",
        "idTipoUsuario",
        "nome",
        "senha"
      ],
      "type": "object",
      "properties": {
        "idUsuario": {
          "format": "uuid",
          "type": "string"
        },
        "nome": {
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "minLength": 1,
          "type": "string"
        },
        "senha": {
          "maxLength": 60,
          "minLength": 5,
          "type": "string"
        },
        "idTipoUsuario": {
          "format": "uuid",
          "type": "string"
        },
        "tipoUsuario": {
          "$ref": "#/definitions/TiposUsuario"
        }
      },
      "additionalProperties": false
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Value: Bearer TokenJWT "
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}